pyproject.toml
[project]
name = "venv-apex-digital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.13"
dependencies = [
    "feature-engine>=1.9.3",
    "fireducks>=1.4.1",
    "loguru>=0.7.3",
    "mlflow>=3.4.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "scalar-fastapi>=1.4.1",
    "scikit-learn>=1.7.2",
    "xgboost>=3.0.5",
]

[dependency-groups]
dev = [
    "empiricaldist>=0.9.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
]
